% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popmunicipal.R
\name{popmunicipal}
\alias{popmunicipal}
\title{Estima populações municipais por faixa etária para o Brasil}
\usage{
popmunicipal(anoin = 2012, anofim = 2020, uano = 2015, infanciadet = F)
}
\arguments{
\item{anoin}{Primeiro ano para obter estimativa}

\item{uano}{Último ano com tabela municipal disponível do IBGE para o TCU com desagregação por faixa etária municipal. Por padrão, 2015.}

\item{infanciadet}{Detalhar a faixa etária de primeira infância. Atualmente funcional até um ano após último ano disponível de dados de nascidos vivos pelo TABNET/DATASUS, isto é ~ano atual -2.}

\item{anofin}{Último ano de que se deseja obter informações}
}
\description{
O objetivo deste script é baixar e processar as estimativas
populacionais do IBGE para construção de indicadores 
nacionais, estaduais ou municipais com os demais
dados municipais.
As bases são o SIDRA e o DATASUS.
}
